support video component

+++ taro/src/storage.ts
@@ -1,4 +1,4 @@
-import { shouldBeObject, getParameterError, temporarilyNotSupport } from './utils'
+import {getParameterError, temporarilyNotSupport} from './utils'
 import { MethodHandler } from './utils/handler'
 
 function getItem(key) {
@@ -17,6 +17,11 @@ function getItem(key) {
 
 // 数据缓存
 export const setStorageSync = (key, data = '') => {
+    if (process.env.NODE_ENV === 'development' && typeof window === 'undefined') {
+        console.error('setStorageSync does nothing on the server-side.')
+        return
+    }
+
     if (typeof key !== 'string') {
         console.error(getParameterError({
             name: 'setStorage',
@@ -37,16 +42,15 @@ export const setStorageSync = (key, data = '') => {
     localStorage.setItem(key, JSON.stringify(obj))
 }
 
-export const setStorage = (options) => {
-    // options must be an Object
-    const isObject = shouldBeObject(options)
-    if (!isObject.flag) {
-        const res = { errMsg: `setStorage:fail ${isObject.msg}` }
-        console.error(res.errMsg)
-        return Promise.reject(res)
+export const setStorage = ({ key, data, success, fail, complete }) => {
+    if (process.env.NODE_ENV === 'development' && typeof window === 'undefined') {
+        fail?.()
+        complete?.()
+        const msg = 'setStorage is always fail on the server-side.'
+        console.error(msg)
+        return Promise.reject(new Error(msg))
     }
 
-    const { key, data, success, fail, complete } = options
     const handle = new MethodHandler({ name: 'setStorage', success, fail, complete })
 
     if (typeof key !== 'string') {
@@ -66,6 +70,11 @@ export const setStorage = (options) => {
 export const revokeBufferURL = temporarilyNotSupport('revokeBufferURL')
 
 export const removeStorageSync = (key: string) => {
+    if (process.env.NODE_ENV === 'development' && typeof window === 'undefined') {
+        console.error('removeStorageSync does nothing on the server-side.')
+        return
+    }
+
     if (typeof key !== 'string') {
         console.error(getParameterError({
             name: 'removeStorage',
@@ -78,15 +87,15 @@ export const removeStorageSync = (key: string) => {
     localStorage.removeItem(key)
 }
 
-export const removeStorage = (options) => {
-    // options must be an Object
-    const isObject = shouldBeObject(options)
-    if (!isObject.flag) {
-        const res = { errMsg: `removeStorage:fail ${isObject.msg}` }
-        console.error(res.errMsg)
-        return Promise.reject(res)
+export const removeStorage = ({ key, success, fail, complete }) => {
+    if (process.env.NODE_ENV === 'development' && typeof window === 'undefined') {
+        fail?.()
+        complete?.()
+        const msg = 'removeStorage is always fail on the server-side.'
+        console.error(msg)
+        return Promise.reject(new Error(msg))
     }
-    const { key, success, fail, complete } = options
+
     const handle = new MethodHandler({ name: 'removeStorage', success, fail, complete })
 
     if (typeof key !== 'string') {
@@ -104,6 +113,11 @@ export const removeStorage = (options) => {
 }
 
 export const getStorageSync = (key) => {
+    if (process.env.NODE_ENV === 'development' && typeof window === 'undefined') {
+        console.error('getStorageSync does nothing on the server-side.')
+        return
+    }
+
     if (typeof key !== 'string') {
         console.error(getParameterError({
             name: 'getStorageSync',
@@ -120,6 +134,11 @@ export const getStorageSync = (key) => {
 }
 
 export const getStorageInfoSync = () => {
+    if (process.env.NODE_ENV === 'development' && typeof window === 'undefined') {
+        console.error('getStorageInfoSync does nothing on the server-side.')
+        return
+    }
+
     const res = {
         keys: Object.keys(localStorage),
         limitSize: NaN,
@@ -129,20 +148,27 @@ export const getStorageInfoSync = () => {
 }
 
 export const getStorageInfo = ({ success, fail, complete } = {} as any) => {
+    if (process.env.NODE_ENV === 'development' && typeof window === 'undefined') {
+        fail?.()
+        complete?.()
+        const msg = 'getStorageInfo is always fail on the server-side.'
+        console.error(msg)
+        return Promise.reject(new Error(msg))
+    }
+
     const handle = new MethodHandler({ name: 'getStorageInfo', success, fail, complete })
     return handle.success(getStorageInfoSync())
 }
 
-export const getStorage = (options) => {
-    // options must be an Object
-    const isObject = shouldBeObject(options)
-    if (!isObject.flag) {
-        const res = { errMsg: `getStorage:fail ${isObject.msg}` }
-        console.error(res.errMsg)
-        return Promise.reject(res)
+export const getStorage = ({ key, success, fail, complete }) => {
+    if (process.env.NODE_ENV === 'development' && typeof window === 'undefined') {
+        fail?.()
+        complete?.()
+        const msg = 'getStorage is always fail on the server-side.'
+        console.error(msg)
+        return Promise.reject(new Error(msg))
     }
 
-    const { key, success, fail, complete } = options
     const handle = new MethodHandler({ name: 'getStorage', success, fail, complete })
 
     if (typeof key !== 'string') {
@@ -168,10 +194,23 @@ export const getStorage = (options) => {
 export const createBufferURL = temporarilyNotSupport('createBufferURL')
 
 export const clearStorageSync = () => {
+    if (process.env.NODE_ENV === 'development' && typeof window === 'undefined') {
+        console.error('clearStorageSync does nothing on the server-side.')
+        return
+    }
+
     localStorage.clear()
 }
 
-export const clearStorage = ({ success, fail, complete } = {} as any) => {
+export const clearStorage = ({ success, fail, complete }) => {
+    if (process.env.NODE_ENV === 'development' && typeof window === 'undefined') {
+        fail?.()
+        complete?.()
+        const msg = 'clearStorage is always fail on the server-side.'
+        console.error(msg)
+        return Promise.reject(new Error(msg))
+    }
+
     const handle = new MethodHandler({ name: 'clearStorage', success, fail, complete })
     clearStorageSync()
     return handle.success()
